refactor: Attendance, AttendanceBook model 패키지로 이동
