+++ src/test/java/AttendanceUpdateTest.java
@@ -1,6 +1,8 @@
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
+import model.Attendance;
+import model.AttendanceBook;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
